Split `test_macos.dart` into `test_ilxd_crypto_bridge.dart` and `test_ilxd_zk_bridge.dart` for better platform compatibility and modular testing.
- Removed `test_macos.dart` and created separate test files for `ilxd_crypto_bridge` and `ilxd_zk_bridge`.
- Updated `build_macos.sh` to run the new test files independently.

This change allows us to focus on getting something running with `ilxd_crypto` on macOS and Android, even though `ilxd_zk` does not yet support macOS in the Rust libraries. By splitting the tests, we can continue work on Android while addressing the Rust project issues for macOS and eventually Android.

Next steps will include creating a `test_mobile.dart` app to consolidate testing across iOS, Android, and macOS, with future expansion to Windows and Linux. The split tests will allow us to easily integrate the same test logic into the Flutter backend for a visual tester interface. This will let us verify that crypto and zk functions (where available) work on the target platform.
