+++ ilxd_bridge/test/test_ilxd_crypto_bridge.dart
@@ -1,70 +1,18 @@
+// test_ilxd_crypto_bridge.dart
 // Copyright (c) 2024 Project Illium
 // This work is licensed under the terms of the MIT License
 // For a copy, see <https://github.com/project-illium/mobilewallet/blob/main/LICENSE>
-import 'dart:io';
+
 import 'dart:typed_data';
-import 'package:ilxd_bridge/ilxd_zk_bridge.dart';
-import 'package:ilxd_bridge/ilxd_crypto_bridge.dart';
 import 'package:crypto/crypto.dart';
-
-int testLoadPublicParams() {
-  int failed = 0;
-  try {
-    print('testLoadPublicParams: Loading public parameters...');  
-    IlxdZkBridge.loadPublicParams();
-    print('testLoadPublicParams: Public parameters loaded successfully.');
-  } catch (e) {
-    print('testLoadPublicParams: Error: $e');
-    failed = 1;
-  }
-  return failed;
-}
-
-int testLurkCommit() {
-  int failed = 0;
-  final expr = '555'; // Example Lurk expression
-
-  try {
-    final result = IlxdZkBridge.lurkCommit(expr);
-    print('testLurkCommit: Lurk commit result: ${result.length} bytes');
-    print('testLurkCommit: Hex representation: ${result.map((b) => b.toRadixString(16).padLeft(2, '0')).join()}');
-  } catch (e) {
-    print('testLurkCommit Error: $e');
-    failed = 1;
-  }
-  return failed;
-}
-
-int testCreateProof() {
-  int failed = 0;
-  String lurkProgram = '''(lambda (priv pub) (letrec ((or (lambda (a b)
-                                                             (eval (cons 'coproc_or (cons a (cons b nil)))))))
-                                                     (= (or 19 15) 31)))''';  
-  String privateParams = "(cons 7 8)";
-  String publicParams = '';
-  int maxSteps = 100;
-  Uint8List proof = Uint8List(32);
-  proof.fillRange(0, 32, 0);
-  Uint8List outputTag = Uint8List(32);
-  outputTag.fillRange(0, 32, 0);
-  Uint8List outputVal = Uint8List(32);
-  outputVal.fillRange(0, 32, 0);
-
-  try {
-    IlxdZkBridge.createProof(lurkProgram, privateParams, publicParams, maxSteps, proof, outputTag, outputVal);
-    print('testCreateProof: success');
-  } catch (e) {
-    print('testCreateProof Error: $e');
-    failed = 1;
-  }
-  return failed;
-}
+import 'package:ilxd_bridge/ilxd_crypto_bridge.dart';
 
 int testGenerateSecretKey() {
   int failed = 0;
   try {
     final secretKey = IlxdCryptoBridge.generateSecretKey();
-    print('testGenerateSecretKey: Generated secret key: ${secretKey.map((b) => b.toRadixString(16).padLeft(2, '0')).join()}');
+    print(
+        'testGenerateSecretKey: Generated secret key: ${secretKey.map((b) => b.toRadixString(16).padLeft(2, '0')).join()}');
   } catch (e) {
     print('testGenerateSecretKey Error: $e');
     failed = 1;
@@ -78,7 +26,8 @@ int testSecretKeyFromSeed() {
 
   try {
     final secretKey = IlxdCryptoBridge.secretKeyFromSeed(seed);
-    print('testSecretKeyFromSeed: Secret key from seed: ${secretKey.map((b) => b.toRadixString(16).padLeft(2, '0')).join()}');
+    print(
+        'testSecretKeyFromSeed: Secret key from seed: ${secretKey.map((b) => b.toRadixString(16).padLeft(2, '0')).join()}');
   } catch (e) {
     print('testSecretKeyFromSeed Error: $e');
     failed = 1;
@@ -92,7 +41,8 @@ int testPrivToPub() {
 
   try {
     final pubKey = IlxdCryptoBridge.privToPub(privKey);
-    print('testPrivToPub: Public key: ${pubKey.map((b) => b.toRadixString(16).padLeft(2, '0')).join()}');
+    print(
+        'testPrivToPub: Public key: ${pubKey.map((b) => b.toRadixString(16).padLeft(2, '0')).join()}');
   } catch (e) {
     print('testPrivToPub: Error: $e');
     failed = 1;
@@ -109,8 +59,10 @@ int testCompressedToFull() {
 
   try {
     IlxdCryptoBridge.compressedToFull(pubKey, outX, outY);
-    print('testCompressedToFull: X-coordinate: ${outX.map((b) => b.toRadixString(16).padLeft(2, '0')).join()}');
-    print('testCompressedToFull: Y-coordinate: ${outY.map((b) => b.toRadixString(16).padLeft(2, '0')).join()}');
+    print(
+        'testCompressedToFull: X-coordinate: ${outX.map((b) => b.toRadixString(16).padLeft(2, '0')).join()}');
+    print(
+        'testCompressedToFull: Y-coordinate: ${outY.map((b) => b.toRadixString(16).padLeft(2, '0')).join()}');
   } catch (e) {
     print('testCompressedToFull: Error: $e');
     failed = 1;
@@ -126,7 +78,8 @@ int testSign() {
 
   try {
     final signature = IlxdCryptoBridge.sign(privKey, messageDigest);
-    print('testSign: Signature: ${signature.map((b) => b.toRadixString(16).padLeft(2, '0')).join()}');
+    print(
+        'testSign: Signature: ${signature.map((b) => b.toRadixString(16).padLeft(2, '0')).join()}');
   } catch (e) {
     print('testSign: Error: $e');
     failed = 1;
@@ -156,9 +109,6 @@ int testVerify() {
 
 void main() async {
   int failed = 0;
-  failed += testLurkCommit();
-  failed += testLoadPublicParams();
-  failed += testCreateProof();
   failed += testGenerateSecretKey();
   failed += testSecretKeyFromSeed();
   failed += testPrivToPub();
@@ -167,8 +117,8 @@ void main() async {
   failed += testVerify();
 
   if (failed == 0) {
-    print('OK: test_macos: ALL TESTS PASSED.');
+    print('OK: test_ilxd_crypto_bridge: ALL TESTS PASSED.');
   } else {
-    print('KO: test_macos: $failed failed tests.');
+    print('KO: test_ilxd_crypto_bridge: $failed failed tests.');
   }
 }
